fixup! separate table title field element
