+++ elements/status.js
@@ -83,7 +83,7 @@ module.exports = function (dat, stats) {
   if (dat.owner && dat.importer) {
     return html`<div>Watching for updates…</div>`
   }
-  var progress = Math.floor(dat.progress * 100)
+  var progress = Math.floor(dat.progress || 0 * 100)
   var progressbarLine = (stats.state === 'loading')
     ? 'line-loading'
     : (stats.state === 'paused' || stats.state === 'stale')
